requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
transformers>=4.30.0
torch>=2.0.0
pydantic>=2.0.0
python-dotenv>=0.19.0
sentence-transformers>=2.2.0 
sentencepiece>=0.2.0
presidio-analyzer>=2.2.358
presidio-anonymizer>=2.2.358
protobuf>=3.0.0
